# doc/api-docs.json
{"openapi":"3.0.1","info":{"title":"USER","version":"v0"},"servers":[{"url":"http://172.20.10.6:52858/user-service","description":"Generated server url"}],"paths":{"/signup":{"post":{"tags":["login-ui-resource-impl"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/signin":{"post":{"tags":["login-ui-resource-impl"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponse"}}}}}}},"/users":{"get":{"tags":["user-ui-resource-impl"],"operationId":"getAllUsers","parameters":[{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"limit","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTUserDto"}}}}}}},"/users/{username}":{"get":{"tags":["user-ui-resource-impl"],"operationId":"getUser","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTUserDto"}}}}}}},"/users/valid/{username}":{"get":{"tags":["user-ui-resource-impl"],"operationId":"validUsername","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}}},"components":{"schemas":{"SignUpRequest":{"required":["email","name","password","username"],"type":"object","properties":{"name":{"maxLength":40,"minLength":4,"type":"string"},"username":{"maxLength":15,"minLength":3,"type":"string"},"email":{"maxLength":40,"minLength":0,"type":"string"},"phone":{"maxLength":20,"minLength":0,"pattern":"^(\\+\\d{3})\\d{5,16}$","type":"string"},"password":{"maxLength":20,"minLength":6,"type":"string"}}},"LoginRequest":{"required":["password","usernameOrEmailOrPhone"],"type":"object","properties":{"usernameOrEmailOrPhone":{"type":"string"},"password":{"type":"string"}}},"JwtAuthenticationResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"tokenType":{"type":"string"}}},"ApiResponseListTUserDto":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TUserDto"}}}},"TUserDto":{"type":"object","properties":{"userId":{"type":"string"},"name":{"type":"string"},"username":{"type":"string"},"phone":{"type":"string"},"status":{"type":"string"},"dtCreated":{"type":"string","format":"date-time"},"email":{"type":"string"}}},"ApiResponseTUserDto":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/TUserDto"}}},"ApiResponseString":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"string"}}}}}}